test: migrate testProp to test.prop API and fix assertions
Swap testProp import alias for the native test.prop syntax from
@fast-check/jest. Reorder arguments to match the new API signature
(arbitraries array first, description second). Fix correctField
lookup in guiBuilder to use index position instead of substring
match, and capitalise 'Everyone' in the no-losers message assertion.
Update fc.float bounds to use Math.fround for accurate float32 range
in validators test.

+++ src/__tests__/unit/validators.test.js
@@ -77,7 +77,7 @@ test.prop([fc.integer({ min: 1, max: 1_000_000 }).map((n) => String(-n))])(
 );
 
 test.prop([
-  fc.float({ min: 0.1, max: 999.9, noNaN: true, noDefaultInfinity: true })
+  fc.float({ min: Math.fround(0.1), max: Math.fround(999.9), noNaN: true, noDefaultInfinity: true })
     .map((n) => n.toFixed(1)),
 ])(
   'validateTimeout rejects decimal strings',
